+++ compiler/src/main/kotlin/dev/zacsweers/metro/compiler/ir/IrContributionVisitor.kt
@@ -44,24 +44,22 @@ internal class IrContributionVisitor(context: IrMetroContext) :
       return
     }
 
-    // Check if it's a plain old ContributesTo
-    for (contributesToAnno in
-      declaration.annotationsIn(symbols.classIds.contributesToLikeAnnotations)) {
-      val scope =
-        contributesToAnno.scopeOrNull()
-          ?: with(metroContext) {
-            diagnosticReporter
-              .at(declaration)
-              .report(
-                MetroIrErrors.METRO_ERROR,
-                "No scope found for @${contributesToAnno.annotationClass.name} annotation",
-              )
-            exitProcessing()
-          }
-      if (declaration.isAnnotatedWithAny(symbols.classIds.bindingContainerAnnotations)) {
+    // @BindingContainer handling
+    if (declaration.isAnnotatedWithAny(symbols.classIds.bindingContainerAnnotations)) {
+      for (contributesToAnno in
+        declaration.annotationsIn(symbols.classIds.contributesToLikeAnnotations)) {
+        val scope =
+          contributesToAnno.scopeOrNull()
+            ?: with(metroContext) {
+              diagnosticReporter
+                .at(declaration)
+                .report(
+                  MetroIrErrors.METRO_ERROR,
+                  "No scope found for @${contributesToAnno.annotationClass.name} annotation",
+                )
+              exitProcessing()
+            }
         data.addBindingContainerContribution(scope, declaration)
-      } else {
-        data.addContribution(scope, declaration.defaultType)
       }
       return
     }